moved separate telemetry prefs page into existing one
