--- src/main/java/com/redhat/devtools/intellij/tektoncd/settings/TektonTelemetryConfigurable.java
@@ -1,51 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021 Red Hat, Inc.
- * Distributed under license by Red Hat, Inc. All rights reserved.
- * This program is made available under the terms of the
- * Eclipse Public License v2.0 which accompanies this distribution,
- * and is available at http://www.eclipse.org/legal/epl-v20.html
- *
- * Contributors:
- * Red Hat, Inc.
- ******************************************************************************/
-package com.redhat.devtools.intellij.tektoncd.settings;
-
-import com.intellij.openapi.options.Configurable;
-import org.jetbrains.annotations.Nls;
-import org.jetbrains.annotations.Nullable;
-
-import javax.swing.JComponent;
-
-public class TektonTelemetryConfigurable implements Configurable {
-
-    private TektonTelemetryView view;
-
-    @Override
-    public @Nls(capitalization = Nls.Capitalization.Title) String getDisplayName() {
-        return "Tekton Pipelines Telemetry";
-    }
-
-    @Override
-    public JComponent createComponent() {
-        this.view = new TektonTelemetryView();
-        return view.getPanel();
-    }
-
-    @Override
-    public boolean isModified() {
-        return false;
-    }
-
-    @Override
-    public void apply() {
-    }
-
-    @Override
-    public void reset() {
-    }
-
-    @Override
-    public void disposeUIResources() {
-        view = null;
-    }
-}
